feat(experiments/cli-args): check exe name
For an ongoing thought experiment I'm doing I wanted to try out a
solution in Clojure to dispatch on binary name (like busybox does) via
the first argument `argv[0]`. It doesn't seem to be available in
Java/Clojure, or else.

In Rust, it's not available under env::args. It is however under
`env::current_exe()`.